requirements.txt
cdlib==0.4.0
matplotlib==3.9.3
networkx==3.4.2
numpy==2.1.3
psutil==6.1.0
python_igraph==0.11.8
scikit_learn==1.5.2
